src/AI/requirements.txt
fastapi
uvicorn
pytesseract
Pillow
python-dotenv
openai
python-multipart
Flask